Add support for tracking metadata blocks in the log.
This patch adds two important data structures, jmap and transaction_infos,
and supporting functions.  Jmap is a map from a metadata block number to
the log block number.  When a transaction commits, jmap is updated with new
mappings; when a block is revoked, the mapping for the block is removed
from the jmap.  Transaction_infos is an array of transaction_info
structures that contain information about transactions currently present in
the log.  It contains a linked list of live blocks in a transaction, and it
is updated after every commit to keep the list up-to-date.
Transaction_infos array will be used by the cleaner for identifying live
blocks and migrating them to appropriate location.

[ Modified by tytso to conditionalize changes on the JBD2_LAZY journal flag ]

Signed-off-by: Abutalib Aghayev <agayev@cs.cmu.edu>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>